# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project Management API",
        "description": "API documentation for the PMS backend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Local API"
        }
    ],
    "paths": {
...
        "/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "List projects",
                "description": "Returns a list of projects in the current workspace.",
                "operationId": "listProjects",
                "parameters": [
                    {
                        "name": "is_archived",
                        "in": "query",
                        "description": "Filter by archived status",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "health",
                        "in": "query",
                        "description": "Filter by health status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "on-track",
                                "at-risk",
                                "off-track"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Projects retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create a new project",
                "description": "Creates a new project in the current workspace.",
                "operationId": "createProject",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "The name of the project",
                                        "type": "string"
                                    },
                                    "key": {
                                        "description": "The project key (e.g., PROJ)",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "The description of the project",
                                        "type": "string"
                                    },
                                    "color": {
                                        "description": "Color for the project",
                                        "type": "string"
                                    },
                                    "team_id": {
                                        "description": "ID of the team",
                                        "type": "integer"
                                    },
                                    "lead_id": {
                                        "description": "ID of the project lead",
                                        "type": "integer"
                                    },
                                    "start_date": {
                                        "description": "Start date",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "target_date": {
                                        "description": "Target date",
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project created successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/projects/{project}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get project details",
                "description": "Returns details of a specific project.",
                "operationId": "getProject",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "The ID of the project",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Update a project",
                "description": "Updates an existing project.",
                "operationId": "updateProject",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "The ID of the project",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "The name of the project",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "The description of the project",
                                        "type": "string"
                                    },
                                    "color": {
                                        "description": "Color for the project",
                                        "type": "string"
                                    },
                                    "health": {
                                        "description": "Health status",
                                        "type": "string",
                                        "enum": [
                                            "on-track",
                                            "at-risk",
                                            "off-track"
                                        ]
                                    },
                                    "progress": {
                                        "description": "Progress percentage (0-100)",
                                        "type": "integer"
                                    },
                                    "team_id": {
                                        "description": "ID of the team",
                                        "type": "integer"
                                    },
                                    "lead_id": {
                                        "description": "ID of the project lead",
                                        "type": "integer"
                                    },
                                    "start_date": {
                                        "description": "Start date",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "target_date": {
                                        "description": "Target date",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "is_archived": {
                                        "description": "Whether the project is archived",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Project updated successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Delete a project",
                "description": "Deletes an existing project.",
                "operationId": "deleteProject",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "The ID of the project",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Project deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/projects/{project}/favorite": {
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Toggle project favorite status",
                "description": "Toggles the favorite status of a project.",
                "operationId": "toggleFavorite",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "The ID of the project",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorite status toggled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_favorite": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/projects/{project}/members": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "List project members",
                "description": "Returns a list of members in the project.",
                "operationId": "listProjectMembers",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "The ID of the project",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Members retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Add member to project",
                "description": "Adds a member to the project.",
                "operationId": "addProjectMember",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "The ID of the project",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID of the user to add",
                                        "type": "integer"
                                    },
                                    "role": {
                                        "description": "Role in the project",
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "member",
                                            "viewer"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Member added to project"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/projects/{project}/members/{user}": {
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Remove member from project",
                "description": "Removes a member from the project.",
                "operationId": "removeProjectMember",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "The ID of the project",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "description": "The ID of the user to remove",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Member removed from project"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}